topdown/json: WIP commit for DAG-based patch algorithm.
This new patch algorithm performs a back-to-front DAG construction
process, which results in an optimal ordering for patch operations
affecting particular keys. By following the "heads" of operation chains
identified by the algorithm, it should be possible to recursively apply
the patch operations in a nearly-optimal fashion. This should mean only
one recursive object construction pass is required, which should give us
enormous reductions in GC cost.

The same approach should be applicable later for the REST API's `PATCH`
method as well.

Signed-off-by: Philip Conrad <philipaconrad@gmail.com>