types/types: Experiment with lazy-sorted Any type.
This commit records the state of the experiment as of 2022-12-14. It
does not improve benchmark results in a worthwhile way, because sorting
is almost immediately required in every instance the type is used.

If there were a way to make `Contains` checks cheaper, and not require a
full sorting, that would solve most of the performance problems.

Signed-off-by: Philip Conrad <philipaconrad@gmail.com>