types: Collapse Any type when >=100 elements.
This commit is a change suggested by @johanfylling, which eagerly
collapses `Any` types in the typechecker to their most general form when
their element counts would exceed 100 elements.

This nets a 10x improvement across all metrics on the
`BenchmarkCompileDynamicPolicy` benchmark, which we've been using as a
proxy for progress on issue #5216.

Signed-off-by: Philip Conrad <philipaconrad@gmail.com>