bundle/file: Optimize loading files from tarballs.
This commit contains a memory-saving optimization, when loading files
from tarballs. We use the header information from the tarball to ensure
Golang does not over-allocate memory for the `bytes.Buffer` we read the
file contents into.

By default, Golang will *double* the capacity of a `bytes.Buffer` every
time it runs out of storage space for bytes being written to it. To
ensure we only allocate what we need, we call `Grow` just once on the
buffer, which saves a few reallocs on larger files, and often results in
a perfectly-sized buffer for the file contents.

Signed-off-by: Philip Conrad <philipaconrad@gmail.com>